+++ src/services/positions/getRolesForStaff/index.ts
@@ -2,7 +2,7 @@ import { AxiosRequestConfig } from "axios";
 import { IRoleForStaff } from "@ptypes/rolesForStaff";
 import { getWithRetries } from "@services/core/getWithRetries";
 import { axiosInstance } from "@api/isaas";
-import { mapRolesStaffApiToEntities } from "./mappers";
+import { mapRolesStaffApiToEntities } from "./mappers/mapRolesStaffApiToEntities";
 
 const getRolesForStaff = async (): Promise<IRoleForStaff[]> => {
   const config: AxiosRequestConfig = {
