+++ src/services/positions/getRolesForStaff/mappers/mapRolesStaffApiToEntities/index.ts
@@ -1,0 +1,10 @@
+import { IRoleForStaff } from "@ptypes/rolesForStaff";
+import { mapRolesStaffApiToEntity } from "../mapRolesStaffApiToEntity";
+
+const mapRolesStaffApiToEntities = (
+  roles: IRoleForStaff[]
+): IRoleForStaff[] => {
+  return roles.map(mapRolesStaffApiToEntity);
+};
+
+export { mapRolesStaffApiToEntities };
